Add a framework for new buffer interface unit testing
Add a new extension type pygame.newbuffer.BufferMixin which allows a Python
new-style class to export a new buffer interface. This enables the creation
of any kind of test array, even broken, in Python. In conjunction with
the ctypes module, it extends testing to multidimensional arrays, without
requiring a third party package such as NumPy.

As an example case, array.array is replaced with a BufferMixin subtype in
bufferproxy_test.py. This was necessary since array.array does not support
the new buffer protocol in Python 2.6. or 2.7.