Fix pygame.pixelcopy function problems with big-endian processors
The pixelarray module was untested on big-endian systems, like the
Power PC. A Mac build of Pygame failed many unit tests. This is a
first attempt to fix the bugs uncovered in pixelcopy. This bug fix
is untested on a Mac.

Also fixed was a faulty unit test in pixelcopy_test.py—a source array
was uninitialized—as well as the bugs uncovered when the unit test was
repaired.