Another bug fix towards issue #175
This adds another missing `Py_DECREF()` for a returned PyNumber_Index object.
In chasing down the lockobj PyPixelArray struct field was removed, with
the `PixelArray` instance directly calling `PySurface_LockBy()` and
`PySurface_UnlockBy()`. The lifetime lock object was redundant and considered a
possible candidate for the leak. Also, the `PyObject_ClearWeakRefs()` call in
`_pxarray_dealloc()` was moved to the to of the slot function, as per the
[Python 3.3 weak reference support documentation](http://docs.python.org/3.3/extending/newtypes.html#weak-reference-support).