reduce pygame.freetype.Font.__init__ args (towards Issue #75)
The keyword args were reduced to a few general font properties. The ptsize
argument is kept for compatibility with pygame.font.Font. The font_index and
resolution arguments are constant properties of the freetype.Font instance.
The ucs4 argument is Python interpreter related property that does not affect
rendering. Omitted properties are either not supported for bitmap fonts
(eg. style with values STYLE_STRONG and STYLE_OBLIQUE), render specific
(eg. style values STYLE_UNDERLINE and STYLE_WIDE, rotation, and vertical),
or obscure and so not likely to be changed (eg. use_bitmap_strikes, strength,
and kerning).

To compiment the ucs4 Font argument, a usc4 property is added. Also, better
ucs4 value checking is performed.