For Surface, Rect, and Color types, resolve Issue #280
For these types, methods or operations that returned a new instance always
return an instance of the base type. For instance, Surface.copy() always
returned a new Surface, event for a subclass of Surface. This changeset
makes Pygame more subclass friendly by altering Surface.copy() and the like
to return a new instance of class of the object on which it is called.

Surface methods changed to work better on subclasses are copy(), convert(),
convert_alpha(), and subsurface(). For Rect the copy(), move(), inflate(),
clamp(), clip(), union(), union_all(), and fit() methods are updated. For
the Color type the correct_gamma() method and +, -. *. //, %, and ~ operators
are upgraded.